+++ core/src/main/java/org/apache/iceberg/SchemaParser.java
@@ -42,7 +42,6 @@ public class SchemaParser {
   private static final String STRUCT = "struct";
   private static final String LIST = "list";
   private static final String MAP = "map";
-  private static final String VARIANT = "variant";
   private static final String FIELDS = "fields";
   private static final String ELEMENT = "element";
   private static final String KEY = "key";
@@ -182,11 +181,7 @@ public class SchemaParser {
 
   private static Type typeFromJson(JsonNode json) {
     if (json.isTextual()) {
-      if (VARIANT.equalsIgnoreCase(json.asText())) {
-        return Types.VariantType.get();
-      }
-
-      return Types.fromPrimitiveString(json.asText());
+      return Types.typeFromTypeString(json.asText());
     } else if (json.isObject()) {
       JsonNode typeObj = json.get(TYPE);
       if (typeObj != null) {
